requirements.txt
pyTelegramBotAPI==4.23.0
google-generativeai==0.8.5
python-dotenv==1.0.1
